Added older EXEs
Added the EXEs I had generated with py2exe. The names dont match since I have since renamed them but I thought people would like them to try the software out. Will need the python27.dll next to them or in the path to run. Not sure python27.dll redistrobution rights and it doesn't need to be uploaded  3 times. Andy has it in the TTD main folder if you need it.